Renamed SDK to DOC
git-svn-id: svn://svn.code.sf.net/p/ifolder/code/trunk@3546 db2431b9-df06-0410-971c-a38ea9a75485

+++ Makefile
@@ -1,6 +1,6 @@
 #
 # The purpose of this Makefile is to simplify the coordinated building
-# of the simias, addressbook, and ifolder projects.
+# of mulitple projects.
 #
 # PREFIX should be set to location where 'make install' should install files.
 # If not already set in the environment, it defaults here to $HOME/stage.
@@ -18,7 +18,6 @@ endif
 all %:
 	@echo PREFIX=$(PREFIX)
 	$(MAKE) -C simias $@
-#	$(MAKE) -C addressbook $@
 	$(MAKE) -C ifolder $@
 	
 #
@@ -30,23 +29,17 @@ register unregister:
 	fi
 
 #
-# Run autogen.sh for all projects, with or without --enable-debug.
-#
-# Short forms of target names:
-# ag = autogen
-# agd = autogen-debug
+# Run autogen.sh for multiple projects, with or without --enable-debug.
 #
 AUTOGEN_CMD       = ./autogen.sh --prefix=$(PREFIX)
 AUTOGEN_DEBUG_CMD = ./autogen.sh --prefix=$(PREFIX) --enable-debug
-AUTOGEN_SDK_CMD   = ./autogen.sh --prefix=$(PREFIX) --with-ndoc-path='c:/Program Files/NDoc/bin/.net-1.1'
+AUTOGEN_DOC_CMD   = ./autogen.sh --prefix=$(PREFIX) --with-ndoc-path='c:/Program Files/NDoc/bin/.net-1.1'
 
 ag autogen:
 	rm -f */config.cache
 	@echo PREFIX=$(PREFIX)
 	cd simias; $(AUTOGEN_CMD)
 	$(MAKE) -C simias install
-#	cd addressbook; $(AUTOGEN_CMD)
-#	$(MAKE) -C addressbook install
 	cd ifolder; $(AUTOGEN_CMD)
 	$(MAKE) -C ifolder install
 
@@ -55,18 +48,17 @@ agd autogen-debug:
 	@echo PREFIX=$(PREFIX)
 	cd simias; $(AUTOGEN_DEBUG_CMD)
 	$(MAKE) -C simias install
-#	cd addressbook; $(AUTOGEN_DEBUG_CMD)
-#	$(MAKE) -C addressbook install
 	cd ifolder; $(AUTOGEN_DEBUG_CMD)
 	$(MAKE) -C ifolder install
 
-agsdk autogen-sdk:
+agdoc autogen-doc:
+	@if ! `uname -o | grep -iq cygwin`; then \
+		echo "currently doc can only be built on Windows"; exit 1; \
+	fi
 	rm -f */config.cache
 	@echo PREFIX=$(PREFIX)
-	cd simias; $(AUTOGEN_SDK_CMD)
+	cd simias; $(AUTOGEN_DOC_CMD)
 	$(MAKE) -C simias install
-#	cd addressbook; $(AUTOGEN_SDK_CMD)
-#	$(MAKE) -C addressbook install
-	cd ifolder; $(AUTOGEN_SDK_CMD)
+	cd ifolder; $(AUTOGEN_DOC_CMD)
 	$(MAKE) -C ifolder install
